fix multiple bugs
FIX - That history wouldn't delete when browser was closed, a site that was set for deletion was navigated away from, and a tab with a site set for deletion was not removing history as expected.

FIX - When opening a new tab, the extension kept the status of the previous tab unless switched to another tab, then back.